requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
weaviate-client==4.5.0
langchain==0.1.7
langchain-openai==0.0.8
langchain-community==0.0.24
python-dotenv==1.0.1
unstructured[pdf,pptx,docx]==0.10.19
o365==2.0.33
psycopg2-binary==2.9.9
